03_ddl: use "char" instead of char for single byte columns
In commit "03_ddl: optimize ddl for mars2" we mentioned that we use char
instead of char(1), the purpose was to use a 8-bit byval type for single
byte columns, however char is not really what we want.  In fact char is
an alias of char(1), so nothing was changed.  We need to use "char"
instead.